What files were added/changed:
- eachTask.js (script for template)
- changes to main.js
- eachTask.html (template for the actual tasks)

Notes:

CHANGES STILL NEED TO BE MADE AND FIXED.

I still need to display the tasks, make sure the template works, and the add event button needs to work.

Events are tasks now, this is due to the collection being called tasks and if we want to change it than we have to make another collection called events.