What files were added/changed: - changes to main.js - changes to home.css - changes to style.css - changes to complete.html - changes to index.html
Notes:

Functionality is almost there but completion button is still not working.

Style changes need to be made as the date and buttons do not mesh well together.

I am aware I made a delete button, we can decide if we want to keep it there or not but I mainly put that there for the events page.

Overall I am still working on things but almost there.